fix droid

+++ Sources/NIO/Socket.swift
@@ -169,7 +169,9 @@ typealias IOVector = iovec
                  controlBytes: UnsafeMutableRawBufferPointer) throws -> IOResult<Int> {
         // Dubious const casts - it should be OK as there is no reason why this should get mutated
         // just bad const declaration below us.
-        var vec = iovec(iov_base: UnsafeMutableRawPointer(mutating: pointer.baseAddress!), iov_len: pointer.count)
+        var vec = iovec(iov_base:
+                            UnsafeMutableRawPointer(mutating: pointer.baseAddress!),
+                        iov_len: UInt(pointer.count))
         let notConstCorrectDestinationPtr = UnsafeMutableRawPointer(mutating: destinationPtr)
 
         return try withUnsafeHandle { handle in
@@ -194,7 +196,7 @@ typealias IOVector = iovec
     /// - throws: An `IOError` if the operation failed.
     func read(pointer: UnsafeMutableRawBufferPointer) throws -> IOResult<Int> {
         return try withUnsafeHandle {
-            try Posix.read(descriptor: $0, pointer: pointer.baseAddress!, size: pointer.count)
+            try Posix.read(descriptor: $0, pointer: pointer.baseAddress!, size: Int(pointer.count))
         }
     }
 
@@ -212,7 +214,7 @@ typealias IOVector = iovec
                  storage: inout sockaddr_storage,
                  storageLen: inout socklen_t,
                  controlBytes: inout UnsafeReceivedControlBytes) throws -> IOResult<Int> {
-        var vec = iovec(iov_base: pointer.baseAddress, iov_len: pointer.count)
+        var vec = iovec(iov_base: pointer.baseAddress, iov_len: UInt(pointer.count))
 
         return try withUnsafeMutablePointer(to: &vec) { vecPtr in
             return try storage.withMutableSockAddr { (sockaddrPtr, _) in
